# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Flipper AI API
  version: 0.1.0
  description: |
    Multi-marketplace flipping tool API. Find undervalued items, analyze flip potential
    with AI, manage opportunities, communicate with sellers, and list items for resale.
  contact:
    name: Axovia AI
    email: axoviaai@gmail.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://flipper-ai.vercel.app/api
    description: Production
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
    Platform:
      type: string
      enum: [CRAIGSLIST, FACEBOOK_MARKETPLACE, EBAY, OFFERUP, MERCARI]
...
    PostingStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, POSTED, FAILED, CANCELLED]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Listing:
      type: object
      properties:
        id: { type: string }
        externalId: { type: string }
        platform: { $ref: '#/components/schemas/Platform' }
        url: { type: string, format: uri }
        title: { type: string }
        description: { type: string, nullable: true }
        askingPrice: { type: number }
        estimatedValue: { type: number, nullable: true }
        valueScore: { type: number, nullable: true }
        condition: { type: string, nullable: true }
        location: { type: string, nullable: true }
        sellerName: { type: string, nullable: true }
        sellerContact: { type: string, nullable: true }
        imageUrls: { type: array, items: { type: string } }
        category: { type: string, nullable: true }
        status: { type: string }
        scrapedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Message:
      type: object
      properties:
        id: { type: string }
        listingId: { type: string }
        direction: { type: string, enum: [INBOUND, OUTBOUND] }
        status: { type: string }
        subject: { type: string, nullable: true }
        body: { type: string }
        sellerName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
security:
  - session: []
paths:
...
  /listings:
    get:
      tags: [Listings]
      summary: List all listings with filters
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - { name: platform, in: query, schema: { $ref: '#/components/schemas/Platform' } }
        - { name: status, in: query, schema: { type: string } }
        - { name: minScore, in: query, schema: { type: number, minimum: 0, maximum: 100 } }
        - { name: location, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: minPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: dateFrom, in: query, schema: { type: string, format: date-time } }
        - { name: dateTo, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Paginated listing results
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings: { type: array, items: { $ref: '#/components/schemas/Listing' } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      tags: [Listings]
      summary: Create a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalId, platform, url, title, askingPrice]
              properties:
                externalId: { type: string }
                platform: { $ref: '#/components/schemas/Platform' }
                url: { type: string, format: uri }
                title: { type: string, maxLength: 500 }
                description: { type: string, maxLength: 10000 }
                askingPrice: { type: number, minimum: 0 }
                condition: { type: string, maxLength: 100 }
                location: { type: string, maxLength: 500 }
                sellerName: { type: string, maxLength: 200 }
                sellerContact: { type: string, maxLength: 500 }
                imageUrls: { type: array, items: { type: string, format: uri }, maxItems: 20 }
                category: { type: string, maxLength: 200 }
                postedAt: { type: string, format: date-time }
      responses:
        '201': { description: Created }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /listings/{id}:
    get:
      tags: [Listings]
      summary: Get a listing by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Listing details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Listing' }
        '404': { description: Not found }
    patch:
      tags: [Listings]
      summary: Update a listing
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                askingPrice: { type: number }
                notes: { type: string }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Listings]
      summary: Delete a listing
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted }
...
  /messages:
    get:
      tags: [Messages]
      summary: List messages
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - { name: direction, in: query, schema: { type: string, enum: [INBOUND, OUTBOUND] } }
        - { name: status, in: query, schema: { type: string } }
        - { name: listingId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, status, direction, sellerName], default: createdAt } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: Paginated messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/Message' } }
                  pagination:
                    type: object
                    properties:
                      total: { type: integer }
                      limit: { type: integer }
                      offset: { type: integer }
                      hasMore: { type: boolean }
    post:
      tags: [Messages]
      summary: Create a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, body]
              properties:
                listingId: { type: string }
                body: { type: string }
                subject: { type: string }
      responses:
        '201': { description: Created }
  /messages/{id}:
    patch:
      tags: [Messages]
      summary: Approve/update a message
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                approved: { type: boolean }
                body: { type: string }
      responses:
        '200': { description: Updated }

  # ── Search Configs ──────────────────────────────────────────────────────
...
  /posting-queue/{id}:
    patch:
      tags: [Posting Queue]
      summary: Update a queued posting
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: '#/components/schemas/PostingStatus' }
                askingPrice: { type: number }
                title: { type: string }
                description: { type: string }
                scheduledAt: { type: string, format: date-time, nullable: true }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Posting Queue]
      summary: Cancel/delete a queued posting
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted }
...